refactor(bienbot-ui): extract optimistic event broadcast switch into useBienBotOptimistic
Moves the large switch(action.type) block that fires pre-API optimistic
events out of handleExecuteAction into a dedicated useBienBotOptimistic
hook, replacing ~237 lines with a single broadcastOptimistic(action, user?._id) call.
